+++ nanoutils/__init__.py
@@ -4,10 +4,11 @@
 
 from .__version__ import __version__
 
-from . import typing_utils, empty, utils, schema
+from . import typing_utils, empty, utils, numpy_utils, schema
 from .typing_utils import *
 from .empty import *
 from .utils import *
+from .numpy_utils import *
 from .schema import (
     Default, Formatter, supports_float, supports_int,
     isinstance_factory, issubclass_factory, import_factory
@@ -15,10 +16,11 @@ from .schema import (
 
 __author__ = 'B. F. van Beek'
 __email__ = 'b.f.van.beek@vu.nl'
-__version__ = __version__
+version_info = VersionInfo.from_str(__version__)
 
 __all__ = []
 __all__ += typing_utils.__all__
 __all__ += empty.__all__
 __all__ += utils.__all__
+__all__ += numpy_utils.__all__
 __all__ += schema.__all__