+++ nanoutils/empty.py
@@ -1,45 +1,58 @@
 """A module with empty (immutable) iterables.
 
-can be used as default arguments for functions.
+Can be used as default arguments for functions.
 
 Index
 -----
 ===================================== =================================================
-:data:`~nanoutils.EMPTY_SEQUENCE`      An empty :class:`~collections.abc.Sequence`.
-:data:`~nanoutils.EMPTY_MAPPING`       An empty :class:`~collections.abc.Mapping`.
+:data:`~nanoutils.EMPTY_CONTAINER`     An empty :class:`~collections.abc.Container`.
 :data:`~nanoutils.EMPTY_COLLECTION`    An empty :class:`~collections.abc.Collection`.
 :data:`~nanoutils.EMPTY_SET`           An empty :class:`~collections.abc.Set`.
+:data:`~nanoutils.EMPTY_SEQUENCE`      An empty :class:`~collections.abc.Sequence`.
+:data:`~nanoutils.EMPTY_MAPPING`       An empty :class:`~collections.abc.Mapping`.
 ===================================== =================================================
 
 API
 ---
 .. currentmodule:: nanoutils
-.. data:: EMPTY_SEQUENCE
-    :value: ()
+.. data:: EMPTY_CONTAINER
+    :type: Container
+    :value: frozenset()
 
-    An empty :class:`~collections.abc.Sequence`.
-
-.. data:: EMPTY_MAPPING
-    :value: mappingproxy({})
-
-    An empty :class:`~collections.abc.Mapping`.
+    An empty :class:`~collections.abc.Container`.
 
 .. data:: EMPTY_COLLECTION
+    :type: Collection
     :value: frozenset()
 
     An empty :class:`~collections.abc.Collection`.
 
 .. data:: EMPTY_SET
+    :type: Set
     :value: frozenset()
 
     An empty :class:`~collections.abc.Set`.
 
-"""  # noqa: E501
+.. data:: EMPTY_SEQUENCE
+    :type: Sequence
+    :value: ()
+
+    An empty :class:`~collections.abc.Sequence`.
+
+.. data:: EMPTY_MAPPING
+    :type: Mapping
+    :value: mappingproxy({})
+
+    An empty :class:`~collections.abc.Mapping`.
+
+"""
 
 from types import MappingProxyType
-from typing import Mapping, Collection, Sequence, AbstractSet
+from typing import Mapping, Collection, Sequence, AbstractSet, Container
 
-__all__ = ['EMPTY_SEQUENCE', 'EMPTY_MAPPING', 'EMPTY_COLLECTION', 'EMPTY_SET']
+__all__ = [
+    'EMPTY_SEQUENCE', 'EMPTY_MAPPING', 'EMPTY_COLLECTION', 'EMPTY_SET', 'EMPTY_CONTAINER'
+]
 
 #: An empty :class:`~collections.abc.Sequence`.
 EMPTY_SEQUENCE: Sequence = ()
@@ -52,3 +65,6 @@ EMPTY_COLLECTION: Collection = frozenset()
 
 #: An empty :class:`~collections.abc.Set`.
 EMPTY_SET: AbstractSet = frozenset()
+
+#: An empty :class:`~collections.abc.Container`.
+EMPTY_CONTAINER: Container = frozenset()